test_good_evtx_chunk() passes
There was a failure that turns out to be a result of the EvtxChunk
struct being larger than the array representing a valid chunk by
the addition of the `records` element at the end of the struct.
Since sizeof() was being used to determine the size for `memcpy()`,
this caused a segfault. This was fixed by calculating size as
`(uintptr_t)&(chunk->records) - (uintptr_t)chunk` instead.
Also substituted `assert()` with an `ASSERT()` macro that
additionally prints a success message if the test is built with
`DEBUG` defined. This helps with tracking down what test crashed
when it results in an error, such as a sigfault, that doesn't
result in an assert fail.
I need to move over to an actual testing framework, honestly.