Set up Check skeleton for use with CMake
Copied over the contents of the Check CMake example, commenting
out the example-specific bits for reference. The current files
ARE NOT ready for building, and most likely will result in an
error on an attempted `cmake .`. That said, it should be trivial
to fit the code from the development branch into this skeleton.