Fit features_basic_cli to the Check skeleton
Moved everything form lib/ to src/ and moved test_data.h to tests/,
but left test.c in root. It will eventually be deleted, and is will
be replaced by the currently empty tests/check_*.c files. Still
have not put proper values into the CMakeList.txt files, so a
build is still not currently possible.