Changed declerations and added typedefs in src/winlog.h
This Included the typedef of many Windows data types, definitions
of a few structs, and changing variable names and types for the
arguments and returns of the API methods. All of this was done in
an attempt to make this library more directly mirror the Windows
API, so that any code written against said API will hopefully
be more easily ported, and Microsoft's documentation should be
more applicable to this library.