Hard-Coded Struct Sizes
Previously I was attempting to calculate the size of structs using
`sizeof(NAME_OF_STRUCT)`
without realizing this would not give me the correct size of all
the members combined. The issue was not evident until I tried to
use the calculated size to determine the location of a chunk in a
log file based off the index of the chunk.

To correct this problem, the size of the structs have been
hard-coded for the time being, and all calls to `sizeof()` have
been replaced with the name of the hard-coded size macros.