Implemented and Tested `chunk_addr()`
Created the "Find_Chunk_In_File" to test if a chunk could be
located within a log file based off its index, testing for
indexes "0" and "1" for now. I would like to add tests
for indexes latery in the file, and also add a way to check
not just that /a/ chunk was found, but that it is the correct
chunk by hard-coding the correct expected offsets.

A new file `check_assertions.h` was also added, intended to contain
custom assertions, mainly binary string comparisons. I have not yet
successfully implemented these assertions, and the newly added
tests use the string comparison assertion `ck_assert_str_eq()`
since `evtx_chunk_magic` is a null-terminated string.

Note: `chunk_addr()` should probably be renamed to
`evtx_chunk_addr()` if it is to remain in the public API of
`evtx.h` to prevent any conflict with other APIs, such as `evt.h`